fix: land recurrence-identifiability promotion and trough-refinement scoring fixes
Commits a body of work that was already documented in CHANGELOG.md's
Unreleased/Fixed section but never actually committed:

- Corrected trough-refinement validation metrics: boundary_set_overlap
  (was mislabelled boundary_set_inclusion -- it measures interval overlap,
  not full inclusion) plus a genuine new truth_set_containment metric;
  renamed the synthetic truth-derived pass1_boundary comparator to
  synthetic_reference_boundary throughout _synthetic.py and
  _trough_refinement_calibration.py so it can no longer be mistaken for
  real pass-1 detector output; peak_changes/duplicate_or_nonmonotonic/
  new_uncomputable now report None ("not evaluated in span harness")
  instead of hardcoded 0/False, since this per-span harness structurally
  cannot measure integration-level properties -- scripts/run_calibration.py
  gains --fixed-trough-policy to revalidate the frozen tuple against this
  corrected scoring without a grid search, and new all_case_coverage/
  all_case_abstention_rate fields keep the resolvable-truth-only and
  every-span denominators from being conflated.
- hydroseason/_report_export.py: trough_boundary_date now always reports
  the actual operational boundary (trough_month) instead of the support
  interval's end date, which differ once a boundary is refined.
- hydroseason/_dynamic_year.py: exports trough_loss_basis; threads
  measurement_tolerance_pp into the refinement call; per-cell dtype-safe
  assignment in _apply_trough_refinement (bulk row assignment refused to
  write a Timestamp into a column that had been all-NaN/float64 so far).
  hydroseason/_catchment.py: threads trough_refinement_policy through
  analyze_catchment.
- hydroseason/_scientific_defaults.py: adds the calibrated (not promoted)
  TROUGH_GEOMETRY_DEFAULTS candidate_for_established_0_3_0.
- hydroseason/_regime.py: clarifies the aseasonal explanation to
  distinguish "genuinely no cycle" from "insufficient evidence to resolve
  one" -- a non-significant test does not itself prove uniform timing.
- docs/migrations/0.2.0-timing-identifiability.md: adds the
  bounds-of-equivalent-evidence paragraph a prior task
  (.superpowers/sdd/task-12-report.md) reported adding but which was
  missing from the actual file, per docs/superpowers/plans/
  2026-09-03-recurrence-identifiability-correction.md's requirement.
- tests/test_package_surface.py: TroughRefinementPolicy is already
  exported from hydroseason/__init__.py; the test's hardcoded allowlist
  had not been updated to match.
- Commits the supporting evidence bundles these fixes and their existing
  CHANGELOG/migration-doc entries already reference: the final scientific
  hardening review bundle, the low-state-refinement endpoint prototype
  findings, the resolution/stress-test/pass2-smoke case studies, and the
  audit script/report they were produced from.

All tests pass (1609 passed, 1 skipped, 2 xfailed) with zero known
failures.

+++ hydroseason/_state_input.py
@@ -58,12 +58,18 @@ def prepare_monthly_extent(
         frame[value_col] = np.where(counts["n_valid"] > 0, 100.0 * counts["n_water"] / counts["n_valid"], np.nan)
         frame["invalid_pct"] = np.where(counts["n_aoi"] > 0, 100.0 * counts["n_invalid"] / counts["n_aoi"], np.nan)
 
-    frame[value_col] = pd.to_numeric(frame[value_col], errors="coerce")
+    # Force float64 even when the source column is all-integer percentages
+    # (e.g. a hand-built [60, 80, 60, ...] fixture with no fractional
+    # values and no count-derived NaN to force an upcast): downstream
+    # refinement code mutates this column in place with NaN and fractional
+    # replacements, which pandas' strict dtype checking refuses to write
+    # into an int64 column.
+    frame[value_col] = pd.to_numeric(frame[value_col], errors="coerce").astype(float)
     if ((frame[value_col] < 0) | (frame[value_col] > 100)).dropna().any():
         raise ValueError("extent_pct must be between 0 and 100.")
     if "invalid_pct" not in frame:
         frame["invalid_pct"] = np.nan
-    frame["invalid_pct"] = pd.to_numeric(frame["invalid_pct"], errors="coerce")
+    frame["invalid_pct"] = pd.to_numeric(frame["invalid_pct"], errors="coerce").astype(float)
     if ((frame["invalid_pct"] < 0) | (frame["invalid_pct"] > 100)).dropna().any():
         raise ValueError("invalid_pct must be between 0 and 100.")
